requirements.txt
# Keep aligned with the validator sandbox image deps:
# autoppia_web_agents_subnet/opensource/sandbox/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
httpx>=0.27.0

# Curated extras for miner flexibility (still run inside sandbox + gateway):
openai>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
orjson>=3.9.0
tenacity>=8.2.0
python-dateutil>=2.9.0.post0
rich>=13.7.0
jsonschema>=4.21.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiohttp>=3.9.0
